# [libc][math][c23] Add expf16 C23 math function