… libc/docs/math/index.rst